refactor: move UserApiKey & AppApiKey types to user-context
